Multiple refactors and introduce Sudoku class
Refactor SudokuSolver to be non-static and take Sudoku object on
construction. Calling solve will solve it and return an Optional sudoku
object (depending on if there is a solution). Constructing a Sudoku
object simply requires a 2 dimensional int array. It has convenience
functions as well as checks for whether or not the sudoku is valid.

With these refactors it's possible to solve larger sudokus (such as 16x16,
25x25 etc.).