Add thermometer rules to SudokuSolver and refactor solve method
Thermometer rules use a simple tree data structure - each node
containing a Cell as data.

The solve method now separates the action of generating the SAT problem
and solving it. The amount of clauses generated is stored as a field for
statistic purposes.

Need to implement generation of SAT clauses from thermometer rules.